merge/upstream

+++ pkgs/build-support/fetchdarcs/builder.sh
@@ -1,5 +1,4 @@
-if [ -e "$NIX_ATTRS_SH_FILE" ]; then . "$NIX_ATTRS_SH_FILE"; elif [ -f .attrs.sh ]; then . .attrs.sh; fi
-source $stdenv/setup
+runHook preFetch
 
 tagtext=""
 tagflags=""
@@ -15,8 +14,25 @@ elif test -n "$context"; then
     tagflags="--context=$context"
 fi
 
-echo "getting $url $partial ${tagtext} into $out"
+# Repository list may contain ?. No glob expansion for that.
+set -o noglob
 
-darcs get --lazy $tagflags "$url" "$out"
-# remove metadata, because it can change
-rm -rf "$out/_darcs"
+success=
+for repository in $repositories; do
+    echo "Trying to clone $repository $tagtext into $out …"
+    if darcs clone --lazy $tagflags "$repository" "$out"; then
+        # remove metadata, because it can change
+        rm -rf "$out/_darcs"
+        success=1
+        break
+    fi
+done
+
+set +o noglob
+
+if [ -z "$success" ]; then
+    echo "Error: couldn’t clone repository from any mirror" 1>&2
+    exit 1
+fi
+
+runHook postFetch
